:truck: Refs #734 - Move confirmation email to be in same tab as confirmation page
